backend/requirements.txt

fastapi
uvicorn
motor
pydantic
python-jose
passlib[bcrypt]
transformers
torch
accelerate
python-dotenv
httpx
